New time formatting functions
